test(mpsc): inline the bounded buffer tests
The file boundary bought nothing over a test module, and the split
hid how little code the unsafe core actually has.
